chore: move repository mocks to a file
