+++ tests/mocks/repositories/transactionRepository.mock.ts
@@ -1,5 +1,5 @@
-import { type ITransactionRepository } from '../../src/repositories/transactionRepository'
-import { Transaction, TransactionType } from '../../src/domain/transaction'
+import { type ITransactionRepository } from '../../../src/repositories/transactionRepository'
+import { Transaction, TransactionType } from '../../../src/domain/transaction'
 
 export const fakeTransactionEntity = new Transaction({
   amount: 1000,
